+++ src/tables/accounts.rs
@@ -145,7 +145,7 @@ impl<'d> Table<'d, Account> {
             };
 
             match txn.commit() {
-                Ok(_) => break,
+                Ok(()) => break,
                 Err(e) => {
                     if !e.as_ref().starts_with("Resource busy:") {
                         return Err(e).context("failed to update entry");