+++ src/time_series.rs
@@ -245,7 +245,7 @@ impl Database {
 
             *series
                 .entry(count_index)
-                .or_insert_with(HashMap::<NaiveDateTime, i64>::new)
+                .or_default()
                 .entry(value)
                 .or_insert(0) += count;
         }